requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
python-jobspy==1.1.43

# Data processing
pandas==2.1.3
numpy==1.26.2

# Snowflake
snowflake-connector-python==3.5.0
snowflake-snowpark-python==1.11.1

# API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Frontend
streamlit==1.28.2
plotly==5.18.0
altair==5.2.0

# Airflow
apache-airflow==2.7.3
apache-airflow-providers-snowflake==5.1.0

# DBT
dbt-core==1.7.2
dbt-snowflake==1.7.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
cryptography==41.0.7
tenacity==8.2.3
